Move sample_coord method to the base Map class
